Sync code changes from go-chi
