+++ distributed-tracing/go-app/internal/config.go
@@ -1,6 +1,6 @@
 // 2019. Mahesh Voleti (mvoleti@vmware.com)
 
-package beachshirts
+package internal
 
 type BeachShirtsConfig struct {
 	Service        string
@@ -41,44 +41,4 @@ func InitGlobalConfig() {
 		SimDelayChance:   0.3333,
 		SimDelayMS:       1000,
 	}
-}
-
-type Server interface {
-	Start() error
-}
-
-type ErrorStatus struct {
-	Error string `json:"error"`
-}
-
-// Shopping
-type Order struct {
-	StyleName string
-	Quantity  int
-}
-
-type OrderStatus struct {
-	OrderId string
-	Status  string
-}
-
-// Styling
-type PackedShirts struct {
-	Shirts []Shirt
-}
-
-type Shirt struct {
-	Style ShirtStyle
-}
-
-type ShirtStyle struct {
-	Name     string
-	ImageUrl string
-}
-
-// Delivery
-type DeliveryStatus struct {
-	OrderNum    string
-	TrackingNum string
-	Status      string
 }
